--- a/.gitmodules
+++ b/.gitmodules
@@ -21,3 +21,6 @@
 [submodule "nano-pow-server"]
 	path = nano-pow-server
 	url = https://github.com/nanocurrency/nano-pow-server.git
+[submodule "flatbuffers"]
+	path = flatbuffers
+	url = https://github.com/google/flatbuffers.git