--- a/ci/actions/linux/install_deps.sh
+++ b/ci/actions/linux/install_deps.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+sudo mkdir -p /etc/docker && echo '{"ipv6":true,"fixed-cidr-v6":"2001:db8:1::/64"}' | sudo tee /etc/docker/daemon.json && sudo service docker restart;
+
+ci/build-docker-image.sh docker/ci/Dockerfile-base nanocurrency/nano-env:base
+ci/build-docker-image.sh docker/ci/Dockerfile-gcc nanocurrency/nano-env:gcc
+ci/build-docker-image.sh docker/ci/Dockerfile-clang nanocurrency/nano-env:clang