--- a/docker/ci/Dockerfile-clang
+++ b/docker/ci/Dockerfile-clang
@@ -15,5 +15,10 @@ RUN update-alternatives --install /usr/bin/c++ c++ /usr/bin/clang++ 100
 ENV BOOST_ROOT=/tmp/boost
 
 ADD util/build_prep/fetch_boost.sh fetch_boost.sh
-
 RUN ./fetch_boost.sh
+
+# workaround to get a path that can be easily passed into cmake for
+# BOOST_STACKTRACE_BACKTRACE_INCLUDE_FILE 
+# see https://www.boost.org/doc/libs/1_70_0/doc/html/stacktrace/configuration_and_build.html#stacktrace.configuration_and_build.f3
+
+RUN ln -s /usr/lib/gcc/x86_64-linux-gnu/5/include/backtrace.h /tmp/backtrace.h