* Update beta network bootstrap weights for v21, cutoff 7M (#2537)

* Cache hash for multiple block->hash () calls (#2536)

* Remove representatives with closed channels (#2530)

* ASAN error with database transaction tracker json serialization (#2538)

* Lower beta network work threshold to 1/64x base (#2540)

* Bounded memory and redesign in the confirmation height processor (#2531)

* Bounded memory and redesign in the confirmation height processor

* Disable frontiers confirmation for test, block doesn't exist during callback in fork resolution

* Fix rpc.confirmation_height_currently_processing test

* Cement blocks below receives not above

* Fixes gcc build (hopefully)

* Store start and end hash in pending write to remove extraneous IO

* Optimise for the case where the top hash is 2 above cemented frontier

* Fix TSAN issues

* Serg comments

* Use cached genesis_hash in CLI --confirmation_height_clear option

* Set accounts_confirmed_info_size to 0 when clearing

* Remove const for prepare_iterated_blocks_for_cementing as confusing

* State blocks sideband upgrade adding is_send/receive/epoch (#2545)

This changes the epoch byte in the sideband to also store these flags in the 3 most significant bits. These can be used to avoid grabbing the previous block to check its balance in some situations after ledger processing.

The upgrade is done in-place, but takes a long time due to having to a random read to retrieve the previous block. A vacuum is done at the end. During the upgrade the ledger grows to almost 2x the ledger size, and up to 3x is required to vacuum as well.

Due to using the most significant bits, the upgrade can safely be stopped and restarted (from the beginning), and no additional versioning was needed.

* Handle legacy confirm_req using the aggregator (#2541)

* Handle legacy confirm_req using the aggregator

* Fix test node.local_votes_cache

* Adjust test to make sure the cache is used

* Use system polls instead

* Poll all nodes, remove outliers, ddos protection & amend RPC response with endpoint (#2521)

* Poll all nodes and remove some metrics from bounds when consolidating

* Update out of date comment

* Formatting

* Fix clang build on actions with long std::tuple

* Allow square brackets in ipv6 address in RPC

* Merge with develop

* Fix ASAN issue

* Gui review comments

* Make parse_address accept v4 and v6 ip addresses

* Incorrect order of arguments

* Use new cached genesis hash

* Move last_telemetry_req to bootstrap_server

* BOOST_1.69 (#2547)

* Probabilistic network packet filter (#2543)

* Extract stream.hpp from blocks.hpp

* Add a probabilistic filter based on direct mapped caches, keyed by 128-bit SipHash.



The filter receives an array of bytes representing a network packet, and is checked for duplicity of the packet. The probability of a false duplicate is marginal, but not zero, and decreases with the size of the filter. The probability of a false non-duplicate is the infinitesimal probability of a 128-bit SipHash collision.

Items are not normally erased. Instead, if a new item is different from the one at the insertion index (digest % capacity), the old item is replaced. There is also a function to erase an element from the filter, if the digest matches it.

The filter state is protected by a mutex, whereas hashing is performed while not holding it.

Uses 1MB of memory for every 64k elements.

* Remove explicit instantiations

* Remove alias

* Optionally set digest in ::apply and add ::clear method directly from a digest

* Encapsulating election::confirmed so it's implementation can vary. Updating relevant tests.

* Add wallet-processed block to work watcher via the block processor (#2548)

* Add wallet-processed node to work watcher via the block processor

* Remove active size check as the request loop may drop one before the check

* guard policy setting by version checks (#2550)

* guard policy setting by version checks greater or equal short explainations of policies being set

* Restore max block processor signature verification batch size (#2546)

* Restore max block processor signature verification batch size

Removed in https://github.com/nanocurrency/nano-node/pull/2279

* Use assert to validate if else condition

* [ASAN] Access node through a weak_ptr on distributed_work dtor (#2554)

* [ASAN] Access node through a weak_ptr on distributed_work dtor

Since there is nothing to force a distributed_work to get destroyed before the node stops, this is a safer way to access the node on the destructor. Fixes some ASAN issues.

* Add a comment explaining the use of node_w

* [ASAN] fix issue in rpc.work_peer_many

* Return created election when adding to active_transactions (#2551)

* This changes the signature of active_transactions::start to match STL container ::insert signatures by returning a pointer to the newly created election and also a bool whether the election creation took place. This allows the call site to retrieve the election that was inserted and determine if it was the initiator of the election creation. Also updated tests that were taking extra steps to retrieve the newly created election.

* Renaming active_transactions::start/add to insert/insert_impl to more closely match the behavior that's expected.

* Separate inactive votes cache from gap cache (#2542)

* Separate inactive votes cache from gap cache
* Move functional changes to new PR

* Inactive votes cache confirmation status (#2553)

* Adding separate confirmation status / bootstrap status for inactive votes cache (it can be different with disabled lazy bootstrap using config bootstrap_fraction_numerator)
* Apply reviews
* Inactive vote cache item can be confirmed with single vote in tests
* Add test to prevent confirmation without quorum

* Add launch flag --disable_block_processor_unchecked_deletion (#2557)

* Disabled test on actions (#2561)

* Add a fake websocket client to test the node websocket server (#2562)

This client keeps connections alive and runs in a separate thread

* Incremental options for ws confirmation subscription (#2566)

Possibility to add or remove accounts in an existing subscription. This is useful for external wallets that can't use the all_local_accounts flag.

* Fix election calling confirm_if_quorum after destruction (#2563)

When insert_inactive_votes_cache confirms the election it deletes itself from the active roots, causing the next confirm_if_quorum call to access freed memory.

* Update --debug_profile_process CLI test (#2564)

* Update --debug_profile_process CLI test
* fix blocks signatures
* support daemon launch options for profiling
* use block count cache

* Apply reviews changes

* Request telemetry data for local node (#2560)

* Disabled UDP by default (#2555)

* Use only the necessary protocols in UPnP (#2571)

* Do not request UPnP for UDP when disabled

* (unrelated) print endline on flag errors

* Buffer drop policies (#2559)

* Buffer drop policies

* Clarify enum comment

* Add GSL-style narrow_cast (#2567)

* Move back timer comments to header (#2572)

* Work version concept (#2569)

* Work version concept

Adds nano::work_version which is passed to all work_generate and work_validate methods (except for tests, ensured with an assert).

The functional changes are quite minimal, as every caller defaults to the only work version available (work_1):

- RPC work_generate, work_validate and block_create can receive an optional "work_version"
- The work websocket outputs the work version

These changes constitute the pipeline necessary for conditional work validation.

* Use nano::to_string to match the work version text

* Work version concept (#2569)

Adds nano::work_version which is passed to all work_generate and work_validate methods (except for tests, ensured with an assert).

The functional changes are quite minimal, as every caller defaults to the only work version available (work_1):

- RPC work_generate, work_validate and block_create can receive an optional "work_version"
- The work websocket outputs the work version

These changes constitute the pipeline necessary for conditional work validation.

* Use a different confirmation height algorithm when ledger is almost fully cemented (#2544)

* Confirmation height algorithm for almost fully cemented ledgers

* Rename variable to make intent clearer

* Count on no confirmation height changes in database outside processor

* Fix timing issue with existing test

* Unnecessary to list confirmation height as a non-locked table in the blockprocessor

* Stein review comments

* Update member order in confirmation_height_processor

* Serg review - TSAN error on initialization, add latch

* Make sure writes are completed before trying to change to another processor

* Add bounded/unbounded counts to stats and check in tests. Add new test too

* Formatting

* Readd conf height as a non-locked table in the blockprocessor needed with RocksDB

* Fix intermittent rpc.confirmation_height_currently_processing test failure

* Update --debug_profile_bootstrap CLI test (#2556)

* Update debug_profile_bootstrap CLI test
* Fix block count check
* Improve logging
* Use cached block counts
* Apply reviews changes

* Add tests for vote_processor (#2574)

* Fix intermittent node telemetry test failures (#2576)

* IPC 2.0 (#2487)

* IPC 2.0

* Use nano::locked for confirmation subscriber vector

* Remove unused local

* Update toml tests, fix some const issues

* Some access permission improvements

* Comments and nano::locked improvements

* Guilherme review feedback: formatting, advanced cmake flags, disallow deny for roles

* Wesley feedback

* Try to please win build on CI

* Add generated file to git ignore

* install api/flatbuffers/* depending on platform

* correct path for api in MacOS app

* add api to docker image



* Log stable filename (#2534)

* Support stable log filename on rotation

* Add to toml test

* warnings: Remove some unused locals and captures (#2583)

* tsan: race in telemetry::ongoing_single_request_cleanup (#2584)

* Check against op aborted on secure rpc acceptor shutdown (#2582)

* [TSAN] confirmation_height.cemented_gap_below_receive test (#2586)

* [TSAN] vote_processor.flush test (#2587)

* friendly backtraces in actions (#2591)

* symlink to backtrace.h in dockerfile Add defines for BOOST_STACKTRACE_BACKTRACE_INCLUDE_FILE update testing scripts to handle clang linking to gcc backtrace.h symlinked location
Continue on Error in ps1 scripts

* use boost 1.70 for clang to allow for

`BOOST_STACKTRACE_BACKTRACE_INCLUDE_FILE` definition revert win actions guards in assert_internal

* Output stacktrace with custom debug assert (#2568)

* Output stacktrace with debug assert

* Error on CI if assert is used

* Remove <cassert> and add debug_assert to platform specific dirs

* Fix CI regular expression

* Missed updating assert for newly added files

* Remove vote cache & generate new vote if election winner is changed (#2585)

* Improve confirmation_solicitor.batches test (#2580)

* prevent self connection
* enable UDP
* add missing deadline

* Add timestamp to telemetry responses (#2573)

* [RocksDB] Limit write locks to necessary tables (#2592)

* Reworking confirmation_height.dependent_election test so it's less prone to race conditions. (#2589)

* Remove dropped_election_cache in preparation for election refactor. (#2590)

* Remove dropped_election_cache in preparation for election refactor.
* Removing unused parameter.

* Fix rpc telemetry test timestamps (#2595)

* Configurable inactive votes cache size (#2579)

* Use C++17 locally, C++14 on CI (#2597)





* Block work version (#2599)

As of this change, work version refers to different work algorithms. Each block type can unequivocally return its work version. A new algorithm will likely need a new state block definition, to avoid ambiguity, although it could be inferred by trying to validate with both algorithms and then setting the version to the one that was valid (if any).

This change allowed removing work_version as a parameter for some work_validate and work_generate overloads. The work watcher can also infer the version from the block it is watching.

* Minimize work validation calls (#2577)

With this PR there should only be one work_validate call per block until it is processed (then more are done to get the difficulty, a future PR will also remove the need for those).

* Start vote generator for changed winner only if voting is enabled (#2593)

* Reduce active mutex locking with election winner details (#2600)

* Reduce active mutex locking with election winner details

* Make sure root is removed before accessing election

* Serg comment on existing code

* Add flag for vote_processor capacity, and tests (#2575)

Slightly limited in testing due to being unable to lock the vote processor loop (on the other hand, the block processor can be paused by starting a write transaction).

- Add node flag vote_processor_capacity, similar to the block processor flags
- Remove check to always process for the test network, the capacity is more than enough
- The representative levels stay the same but are now calculated on the fly from the given capacity
- ::vote() now returns a boolean representing if the vote was processed, both for testing purposes and will be useful for the upcoming network filter

* Don't use active transactions mutex lock during confirmation solicitor.prepare () (#2602)

* Aggressive flooding for local blocks (#2549)

* Aggressive flooding for local blocks

Floods locally produced blocks (going through the work watcher, or process_local - wallet, RPC process) such that they are sent to all PRs and random subset of other peers.

This allows more protection against Sybil attacks, and more streamlined elections as there are less occasions of votes arriving before blocks.

As a consequence, the average amount of echoes will increase by 1 for PRs. However, with this change in place and used by most of the network, a future enhancement would be to reduce the republishing fanout.

This change is based on a proposal by @Srayman for more information see the original proposal at https://medium.com/nanocurrency/proposal-for-nano-node-network-optimizations-21003e79cdba

* Add flag to disable republishing in block processor and a test for aggressive flooding

* Adjust timings for sanitizer builds

* Wrap in a lambda to simplify test, and adjust it

* Block difficulty and work validation cleanup (#2601)

Blocks now have a method to calculate and return the difficulty based on their work value, work version, and root. Work validation has been cleaned up, removing optional output difficulty from all methods, and with new work_difficulty and work_threshold methods. These are now used where they make sense, such as when previously work_validate was done even though we only wanted to compare the resulting difficulty. Functionally the same, but much less verbose.

Without changing the interface, we can decide to cache the difficulty in the block in the future, if it ever shows up in profiling.

Further changes will be required when a new work version is added, especially if the proof is larger than 8 bytes, but these changes are a step forward in that directtion.

* Bootstrap attempts and connections/pulls separation (#2499)

- New class `nano::bootstrap_connections` to manage client connections & bulk pulls
- Separate source file for connections
- Restored bootstrap clients list (to close connections with `stop` command)
- Parent class `nano::bootstrap_attempt` & child classes for legacy, lazy & wallet bootstraps
- Separate source files for bootstrap attempts & lazy attempts
- Allowing several concurrent bootstrap attempts (currently for different bootstrap modes, in the future can be easily modified to allow same concurrent modes)
- Separate `nano::bootstrap_attempts` class for fast attempt search with incremental ID
- Bulk pull info is modified to include bootstrap attempt incremental ID
- `force` option in bootstrap RPCs & RPC "bootstrap" are currenlty designed to close all concurrent attempts
- Config field `bootstrap_initiator_threads` to manage bootstrap concurrency. Default 2 for multithreaded systems, 1 for singlethreaded & test network
- RPC "bootstrap_status" is modified to show all bootstrap attempts & new connections class
- Reduced lazy bootstrap attempts memory consumption: processed blocks unordered map is modified to store only 64 bit hash of block instead of full 256 bit union
- Fixed `websocket.bootstrap_exited` test TSAN warnings

* Utility: nano::optional_ptr (#2605)

* optional_ptr

* Implement feedback from Wesley/Guilherme

* CMakeLists backwards compatibility (#2607)

use version checks supported on older cmakes
use atleast 3.11 FindBoost.cmake as import targets are not properly exposed on previous versions

* Lock before stopping when it is necessary to notify other threads (#2608)

This prevents having all threads waiting and the system freezing. Mostly important for tests, especially with sanitizers, but could happen on stopping the node normally.

Portmapping requires an atomic bool, and the database queue doesn't after this change.

Other small related bootstrap changes included in this PR per Serg's suggestion.

* Simplify telemetry data processing (#2598)

* Simplify telemetry data processing

* Stein review comments

* Fix network.replace_port

* Launch flag --allow_bootstrap_peers_duplicates (#2606)

Allowing multiple connections to same bootstrap server (useful in tests or for fastest beta nodes sync)

* Fix intermittent send_node_id_handshake unit test failures (#2612)

* Update preconfigured_peers comment regarding port (#2616)

* update url for sourceforge (#2617)

* Election state refactor (#2535)

Converting the election class into a state machine instead of working on rebroadcast iterations.

* Allow CLI --config values for inactive node tests (#2594)

* Attach sideband to block (#2596)

* Move epoch from secure to nano_lib

* Attach sideband to block and always (de)serialize along with it

Using the sideband is only valid for blocks that were processed with code `progress`, otherwise they may not be set (important examples: old, fork).

* Make sideband optional to ensure usage correctness

* Interim

* Use new nano::optional_ptr to hold the sideband

* Adjust some tests to ensure two nodes don't simultaneously process the same block object

* Update comment

* Unchecked deletion (#2609)

* Host qt assets for windows on s3 (#2622)

* [RocksDB] Tests not reading account count from store correctly (#2623)

* Fix confirmation_height.gap_live intermittent test failure (#2621)

* Simplify request aggregator mutex lock behavior (#2614)

* Use the sideband when available in ledger.is_send (#2620)

* confirmation_height.modified_chain test fails on a non-debug build (#2624)

* tsan fix: rpc.wallet_destroy (#2615)

* Improve --debug_profile_bootstrap performance (#2626)

block sideband is not required for some CLI tests

* Election refactor follow up (#2619)

* active_transactions include cleanup

* Allocate a new solicitor on every confirmation request loop

* Increment election confirmation_request_count

…